bitnet/requirements.txt
bitnet==0.2.4
scikit_learn==1.5.2
matplotlib==3.9.2
tqdm==4.66.2
torchinfo==1.8.0
torchmetrics==1.4.3
wandb==0.18.5